backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
azure-storage-blob==12.19.0
cryptography==41.0.7
Pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
drf-spectacular==0.27.0